scripts/pyproject.toml
[project]
name = "profile-scorer-scripts"
version = "0.1.0"
description = "Python scripts for DB inspection and plotting"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "python-dotenv>=1.1.0",
    "wordcloud>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=9.0.0",
]

[tool.uv]
dev-dependencies = [
    "ipython>=9.0.0",
]
